+++ plugin/controllers/models/config.py
@@ -1,9 +1,13 @@
+# -*- coding: utf-8 -*-
+
 from enigma import eEnv
 from Components.SystemInfo import SystemInfo
 from Components.config import config
 from Tools.Directories import resolveFilename, SCOPE_CURRENT_PLUGIN, fileExists
 from os import path, listdir
 import xml.etree.cElementTree
+
+from Plugins.Extensions.OpenWebif.__init__ import _
 
 def addCollapsedMenu(name):
 	tags = config.OpenWebif.webcache.collapsedmenus.value.split("|")
@@ -66,13 +70,15 @@ def getJsonFromConfig(cnf):
 		if type(cnf.choices.choices) == dict:
 			choices = []
 			for choice in cnf.choices.choices:
-				choices.append((choice, cnf.choices.choices[choice]))
+				choices.append((choice, _(cnf.choices.choices[choice])))
 		elif type(cnf.choices.choices[0]) == tuple:
-			choices = cnf.choices.choices
+			choices = []
+			for choice_tuple in cnf.choices.choices:
+				choices.append((choice_tuple[0], _(choice_tuple[1])))
 		else:
 			choices = []
 			for choice in cnf.choices.choices:
-				choices.append((choice, choice))
+				choices.append((choice, _(choice)))
 				
 		return {
 			"result": True,
@@ -169,7 +175,7 @@ def getConfigs(key):
 		for entry in config_entries:
 			try:
 				data = getJsonFromConfig(eval(entry.text or ""))
-				text = entry.get("text", "")
+				text = _(entry.get("text", ""))
 				if "limits" in data:
 					text = "%s (%d - %d)" % (text, data["limits"][0], data["limits"][1])
 				configs.append({
@@ -216,7 +222,7 @@ class ConfigFiles:
 		self.setupfiles = []
 		self.sections = []
 		self.section_config = {}
-		self.allowedsections = ["usage", "recording", "subtitlesetup", "autolanguagesetup", "avsetup", "harddisk", "keyboard", "timezone", "time", "osdsetup", "epgsetup", "lcd", "remotesetup", "softcamsetup", "logs", "timeshift"]
+		self.allowedsections = ["usage", "userinterface", "recording", "subtitlesetup", "autolanguagesetup", "avsetup", "harddisk", "keyboard", "timezone", "time", "osdsetup", "epgsetup", "display", "remotesetup", "softcamsetup", "logs", "timeshift", "channelselection", "epgsettings", "softwareupdate", "pluginbrowsersetup"]
 		self.getConfigFiles()
 
 	def getConfigFiles(self):
@@ -241,6 +247,9 @@ class ConfigFiles:
 			xmldata = setupdom.getroot()
 			for section in xmldata.findall("setup"):
 				configs = []
+				requires = section.get("requires")
+				if requires and not SystemInfo.get(requires, False):
+					continue;
 				key = section.get("key")
 				if key not in self.allowedsections:
 					showOpenWebIF = section.get("showOpenWebIF")
@@ -261,9 +270,9 @@ class ConfigFiles:
 				if len(configs):
 					sections.append({
 						"key": key,
-						"description": section.get("title")
+						"description": _(section.get("title"))
 					})
-					title = section.get("title", "")
+					title = _(section.get("title", ""))
 					self.section_config[key] = (title, configs)
 		sections = sorted(sections, key=lambda k: k['description'])
 		self.sections = sections